+++ Cargo.toml
@@ -130,6 +130,7 @@ lb-zksign                        = { default-features = false, package = "logos-
 blake2            = { default-features = false, version = "0.10" }
 bytes             = { default-features = false, version = "1.3" }
 cached            = { default-features = false, version = "0.55.1" }
+cfg-if            = { default-features = false, version = "1.0.4" }
 ed25519-dalek     = { default-features = false, version = "2" }
 fork_stream       = { default-features = false, version = "0.1.0" }
 hex               = { default-features = false, version = "0.4" }
